添加参数检查

+++ src/mysql.js
@@ -42,6 +42,7 @@ function query (queryString) {
  */
 function insert (tableName, column, value) {
   return new Promise((resolve, reject) => {
+    if (!tableName && !column&&!value) return reject(new Error('tableName , column and value must not be null'))
     let sqlString = mysql.format('INSERT INTO ?? (??)VALUES (?);', [tableName, column, value])
     // console.log(sqlString)
     pool.query(sqlString, function (error, results, fields) {
@@ -61,6 +62,7 @@ function insert (tableName, column, value) {
 function sqlDelete (tableName, whereList) {
   // DELETE FROM 表名称 WHERE 列名称 = 值
   return new Promise((resolve, reject) => {
+    if (!tableName && !whereList) return reject(new Error('tableName and whereList must not be null'))
     let sqlString = 'DELETE FROM ?? '
     if (whereList) {
       if (whereList % 2) return reject(new Error('whereList should be even.'))
